Updated openstack/openstack Project: openstack/nova  aed060a91b3d748beddb302c217264919c8ac92c
Update obj_reset_changes signatures to match

The base object obj_reset_changes method has a recursive parameter that
dives into objects within an object to reset changes.  The Flavor and
Instance objects override that method but are missing that parameter.
The recursive parameter was not added to those objects when it was
changed on the base object.

Closes-bug: #1487522
Change-Id: I7be4cfa59187fffd88413a5ba97b13ff3f93eefc