Updated openstack/openstack Project: openstack/nova  4c9fc8b418dbba747aa87a984589bc76f76bd95b
Handle NotFound exceptions while processing network-changed events

We could be getting a network-changed event from neutron in the case
that something is being torn down on the networking side, in which case
we're in a race on the nova side since we're probably deleting the
instance and it's network info_cache.

We should handle NotFound exceptions in that case so that we don't stop
looping over all events and fail to process any others that come after
that network-changed event.

Closes-Bug: #1487025

Change-Id: Icbc4ed9776de710c0102cf411f5c9808c1a05e83