Updated openstack/openstack Project: openstack/designate  0d3b81c338a75b470b52607c8daed4ae2a110153
Refactor notify logic into backends

The idea here is, like many things, the time and frequency of sending
NOTIFYs is not the same for every backend. So we let the backends decide
when to send their NOTIFYs. Ideally this makes the Pool Manager code simpler
and the backend code easier to understand, and avoids sending some
unnecessary NOTIFYs.

This requires a change to MiniDNS' rpcapi to make notify_zone_changed and
get_serial_number take a host and port, instead of a nameserver object.

Change-Id: I8d67cebeecc8261e715849aefeafd20f0d60d304