Updated openstack/openstack Project: openstack/trove  7cf297cd221f71b5ffae370a23ab3aafc0c15f76
Associate flavor types with datastore versions

This feature provides the ability to associate flavor
types with datastore versions.The trove-manage util
will provide the ability to add or delete this information.
All nova flavors are permitted for a datastore-version unless
one or more entries are found in datastore_version_metadata,
in which case only those are permitted.

partially implements blueprint associate-flavors-datastores

Change-Id: Ib012401f89d07c502f93d5ee1cd4abb6b738953a