Updated openstack/openstack Project: openstack/keystone  40ecf5e61e2d6277d38d5b0bf04201db4f58583b
Add testcase to test invalid region id in request

Regions use ID differently. The user can specify the ID or it will be
generated automatically. So, there is schema validation that region ID
should be the string type, these testcases test where region id could
be passed in request body but not a string type.

Change-Id: I0623e8e97132b9c2e60d03fca8fa98de798b5494
Closes-Bug: #1487663