Updated openstack/openstack Project: openstack/nova  db477bb6f29dac5709b7869ea427f2ef5ea68a5f
Pull the all_tenants search_opts checking code into a common utility

I've always hated the logic involved in the servers extension to check
the all_tenants flag in search_opts and how it doesn't just store a
local boolean variable and re-use it.

This pulls that code out into a common utility method so both the v2 and
v2.1 extensions can use it (rather than duplicate the same mess), and it
cleans up the logic for setting the flag so it makes a bit more sense.
It also decouples the logic that figures out if we're getting all
tenants from the code that removes the entry from the search_opts dict,
so that's left up to the caller.

Change-Id: If5a570223267bf8a737a33398ddbf12d712a7b96