Updated openstack/openstack Project: openstack-infra/project-config  2a4809b9ae05967db947732c86b0c08359131ddc
Add new repository for stand-alone TOSCA Parser

The heat-translator project currently contains two components - Translator and
TOSCA Parser. In order to use them independently we are splitting them to
their own repository so that they can be offered as a PyPI package. The patch
for modifying the governance file is https://review.openstack.org/#/c/211321/
We have decided to start with creating a new repository by copying the existing
repository, git://github.com/openstack/heat-translator.git, and then manually
delete few directories and files that are not required for the stand-alone
Parser.

Change-Id: Ie9f046d3bacf9fe3825577e1068d36dfbba46613