Updated openstack/openstack Project: openstack/python-cinderclient  8b4a5bf6f4b712eacb1f828d401931d4e6af31c1
Show backup and volume info in backup_restore

When no volume is specified in backup_restore command, a new
volume is created with specified name.

If running the same command several times, several volumes
with same name are created, it needs to check logs to find out
which volume is latest.

This update is to show the returned info from api.

Change-Id: I035fc85aa636b716b16c8a543c843ebf4659a96e
Closes-Bug: #1472493
Depends-On: I9f5d8ded993fcf1d351d4c65b6806929693b15d6