Updated openstack/openstack Project: openstack/cinder  6082c1b14857a66996b0645b610faec9cd17cdfa
Add version columns to services table

The following patch is part of the cinder effort to
support rolling upgrade.  This patch adds columns
to the services table to track the RPC and
oslo_versionedobjects versions of each service.

Follow up patches will be made to have each service:
register its RPC and oslo_versionedobjects versions on
startup, make the RPC and oslo_versionedobjects versions
compatible with an older release, and update the versions
once all services are updated to the latest release.

Change-Id: Ifa6c6ac230988c75dcc4e5fe220bfc5ee70ac338
Partial-Implements: blueprint rpc-object-compatibility