Updated openstack/openstack Project: openstack/nova  db7887b95666fccb8f5758ec8ca55494f091725c
Merge virtual interface tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges virtual interface functional tests.
Previously there was no functional tests of virtual interface in v2.1.

There is difference between v2 and v2.1 as extension "OS-EXT-VIF-NET"
is missing in v2.1.
As v2,1 has been released in Kilo, it is too late to fix that difference
in v2.1 base. That fix will go with microversion -
I3caca80a6c010b86150909126f4545425ed99e11
Due to that sample files are different for v2 and v2.1.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled to run
merged functional tests for both v2 and v2.1.

Partially implements blueprint test-collapse-v2-and-v21

Change-Id: I7097483cb57ed785b01fddb4f832e453f71987fd