Updated openstack/openstack Project: openstack/oslo.versionedobjects  aba1089faba22188d3f8f834db7475b6d3312c93
Save and re-raise exception

When an exception is raised, we do "some stuff" and re-raise that
original exception. This uses oslo_util's
excutils.save_and_reraise_exception to do this, to handle the case
where "some stuff" might cause the exception context to be cleared
(which would have resulted in None being attempted to be re-raised).

Change-Id: I91234797ab71399f6f9bac3a3ab35a6eeb2a87f3