Updated openstack/openstack Project: openstack-infra/elastic-recheck  cba66dbda09365d03e72c6ca3f2fc6de21cd4a45
Add config flags for data source configuration

This commit adds options to the config file for the elastic recheck
bot configuration file. This enables users to specify how to connect
to an elastic recheck server and a subunit2sql database, but things
will still default to using the openstack-infra servers to prevent
breaking the running service.

Change-Id: I10db1a568cc01e137e5f4d8a8814b17201c4c438