Updated openstack/openstack Project: openstack/openstack-manuals  603ba41c8131c49c793c0009344aed0c6b323535
Fix double index entry

We cannot use ..only.. with a toctree, it gets parsed at the wrong
time.

Add a second Debian specific index file so that Debian gets only this
file. This avoids the double index entry in the sidetable but leads to
no index table for Debian.

Change-Id: I6f4e7dffb2c5cd4416607817dcb3685b65b0674d
Partial-Fix: #1488187