Updated openstack/openstack Project: openstack/horizon  a80e63f14cadc52f730d6e3c6e1971dc538cf735
Stop using naturalSort in hypervisors table

In some situations, JS's naturalSort sorts hypervisor_hostname in
an unusual order. For example, this is a usually ordered list:

  east-01, east-02, west-01, west-02

while the result made by naturalSort is:

  east-01, west-01, east-02, west-02

Also, this is inconsistent with the other tables as they are using
datatable's default sort. Actually, the default sort does natural
ordering to numbers in text. So we can use the default sort instead
of naturalSort without a problem.

As the hypervisor table is the only user of naturalSort, all stuff
for naturalSort were removed.

Change-Id: I8df3ff72a33d194a5b6bc1692aac7878a611ae44
Closes-Bug: #1480563