Updated openstack/openstack Project: openstack/heat-translator  3f98139a7eb4ab896ad02cdeb7b54b0f5b9f881e
Allow for dynamic class loading of target types

Tosca types need to be mapped to corresponding target translation
types (i.e. HOT). This commit allows for the target types to be
dynamically loaded from a directory. It allows for more flexibility
by pulling types from a known directory instead of defining each
individiual type statically in the code. For example, with this
commit you no longer need a separate import for each target type.

Also, this commit adds the notion of a global configuration object
that can be shared throughout the translator. The configuration
defines the location directory for custom defined target types.
This configuration can be set by the user in the corresponding
translator/conf/translator.conf file. In the future, additional
values can be added to this configuration and the code can be
extended to support them, but only the required values were
implemented here.

Change-Id: If7b8da12eef5b8ed8a2e11b1f412203d4ed59c5a
Implements: blueprint dynamic-tosca-to-hot-map