Updated openstack/openstack Project: openstack/horizon  d8bbdd19d4b430fe3d7a7855f22edbc50f8d4541
Reduce the chance of Selenium timeouts in integration tests

Avoid situations when we trying to take something out of a destroyed
box (contents of an element that is no longer there). Plus other
improvements to a couple of element getter functions/predicates that
might had been causing timeouts.

Related-Bug: #1467950
Implements blueprint: integration-tests-hardening

Change-Id: Ia17cb2134eaa5dd14a04384d1b703a8a648f952d