Updated openstack/openstack Project: openstack/nova  6efcd010e6e29b1e64d64752fbb56b05b7134e5e
Nil out inst.host and inst.node when build fails

This patch addresses a bug that leaves instance.host/node set after
a VM fails to build or reschedule (e.g., due to an error in the spawn
path). Rather, we should nil these out at the end of the build
process so that the instance is not erroneously being associated
with stale host/node information.

Also included are a few cosmetic updates in comments and eliminating
some implied 'None' parameters from "dict.get(key, None)" calls I
found during the course of investigating this change.

Change-Id: I4d9c90b1d3fcce5237323fffcc6db1fb93d23e37
Closes-Bug: #1427944