Updated openstack/openstack Project: openstack/neutron  89a4158444b60349dd95d233ac2cb1a2cad4d7b0
ip_lib: support creating Linux dummy interface

This is for use by a DHCP agent interface driver in the
networking-calico project.  networking-calico connects VMs without
using bridging, so it needs an unbridged DHCP port interface with
which it can associate the DHCP subnet prefix, and the Linux dummy
interface (in conjunction with use of Dnsmasq's --bridge-interface
feature) is suitable for that purpose.

Partial-Bug: #1486649
Change-Id: I5485c187bc44bac3c2942c4fc9e5a502912a81a2