Updated openstack/openstack Project: openstack/swift  3ce602d7fbfe9c68b4121f38351d1869b91135d2
Quorum on durable response is too low

Increase the .durable quorum from 2 to "parity + 1" to guarantee
that we will never fail to rebuild an object.  Otherwise, with
low durable responses back (< parity + 1), the putter objects
return with failed attribute set to true, thereby failing the
rebuild of fragments for an object.

Change-Id: I80d666f61273e589d0990baa78fd657b3470785d
Closes-Bug: 1484565