Updated openstack/openstack Project: openstack-dev/devstack  5bb5be4b450364a7a4ef35286d73092f7012e14b
do not redefine path in elasticsearch

in I0272d56bc2e50e8174db78bd449f65f60f7f4000, we reset DEST value
when installing elasticsearch. it gets set to /opt/stack/ which is
not always correct in gate causing the path to be wrong and
elasticseach cannot be installed. we should reuse DEST from stackrc

Change-Id: Ia3a2383ada30c4e92c37386aedd6164c69cac60a
Closes-Bug: #1484182