Updated openstack/openstack Project: openstack/heat  a8efc2fa79099f1369f758ba14dea0e8124f5094
SoftwareDeployments modify validation to use properties

Currently we rely on an internal implementation detail of the
server resource to validate, which is both potentially fragile
and also inflexible should you wish to override the server
resource with a nested stack resource.

Instead, check the expected value exists within the properties
of the resource, which decouples the two resource implementations
and potentially enables easier overriding of the server resource,
e.g with a nested stack for testing.

Change-Id: Id666f7ae951b4ed31930f09fe0c1b0d8c3fe09ed