Updated openstack/openstack Project: openstack/nova  9d8f84685df7748e9d3d808f781a0dd3510e37ff
Move CPU and RAM allocation ratios to ResourceTracker

In order to have a per-compute allocation ratio basis, we need to provide those
to the compute nodes, so let's provide them there and import them on the
scheduler side until the work is fully completed.

That change specifically implies that nova.conf is consistent across the whole
cloud since what was formerly defined on the scheduler side will have its
definition by the compute nodes.

Release notes for Liberty are already updated to mention the above.

Change-Id: Ic2a80b3804fc27f655e6e74c30f09cab59b16ccb
Partially-Implements: blueprint allocation-ratio-to-resource-tracker