Updated openstack/openstack Project: openstack/nova  936b0127a806dd51586fff8c83f181d60980dbfc
libvirt: refactor _create_domain_setup_lxc to use Image.get_model

Instead of trying to figure out what the proper ImageModel is for
the instance's image, just use the get_model function. The resulting
ImageModel will actually represent the proper ImageModel for the
instance, and downstream mounting functions will be able to make
intelligent decisions on how to mount the instance's disks.

It is not possible to use get_model for an instance booted from volume,
but we can still use LocalBlockImage in it's case because it is
simply exposed as a block device.

Partial-Bug: #1287754
Change-Id: I0f4456731c9d6f64affbf6924f4a7a4c34df4fd1