Updated openstack/openstack Project: openstack/horizon  ab659d20e0f973b7395cc5cce2963f45967429c4
Refactor angular cloud services utilities

A nova-extensions and settings-service services and directives
were implemented at the very end of Kilo.

The directive implementation required a dependency tree that
should have to be maintained.  This removes that dependency
tree. It also greatly improves the ability to customize
the directives, while improving readability, by using the
angular.extend way of creating directives from a template.

Ultimately, this leads to much better decoupling of the
functionality which makes it easier for future plugins
and deployers to write their own directives with similar
functionality.

This is the recommended way by angular docs.
See the bottom of this page for more on extending directives:

  https://github.com/angular/angular.js/wiki/Understanding-Directives

This change also enables the directive to be attribute based,
rather than element based, making it more flexible in usage.

In addition, this takes advantage of compilation that improves
performance by ensuring that non-enabled features are not
compiled, meaning any nested directives will not be processed.

To manually test this with Launch Instance NG:

hz-settings-toggle:
  In local_settings.py the following between True and False
     'can_set_mount_point': True,

  On the first step of Launch Instance, notice that when you select
  Create New Volume that when the above is True, that you'll have
  and input box for setting the device name.  When it is false,
  you won't have the input box.

hz-nova-extensions:
  In openstack_dashboard/dashboards/project/static/dashboard/
     project/workflow/launch-instance/configuration/configuration.html

  Find <div hz-nova-extensions='["ConfigDrive"]'>
  Change to <div hz-nova-extensions='["NON_EXISTENT_EXTENSION"]'>

  Launch the Launch Instance NG. Go to the Configuration step.
  Note that when the extension is ConfigDrive, you'll see the
  checkbox for config drive.  When you set NON_EXISTENT_EXTENSION,
  the checkbox won't exist.

Change-Id: I56b52288d18a2558d2a60878c88714d2672bf783
Partial-Bug: #1450150
Co-Authored-By: Rajat Vig <rajatv@thoughtworks.com>