Updated openstack/openstack Project: openstack/cinder  25d90467b176463b609834ceb5851920896c6975
Move import and export backup metadata to object

Encoding and decoding of backup metadata was being done in the backup
driver, but since this is not driver specific data we are talking about
it makes more sense to have this logic in the backup object.

This patch moves the logic to the object while still allowing to have
importing/exporting driver specific backup data, even when this is not
being used at the moment.

Related-Bug: #1455043
Related-Bug: #1476416
Change-Id: I690748e4542880d5e926f3cbf8cedfc86de5cb1f