Updated openstack/openstack Project: openstack-infra/project-config  50c1386a9dee33948938614e71ac150f825b0c2a
puppet: run openstack-integration jobs for keystone & oslib modules

puppet-openstack-integration jobs are now working:
* Keystone is deployed in WSGI with Puppet modules (keystone &
  openstacklib are the two biggest requirements)
* Tempest is run to validate Keystone v2 and v3 APIs.

Now, we could like to move forward and add the jobs for puppet-keystone
and puppet-openstacklib so we can see how it behaves with Puppet
modules.

In short-term, we would like to get the feedback from the jobs and see if
it needs more adjustment.
In medium-term, we might want to add the jobs to other modules
(puppet-nova, etc).
In long-term, openstack-integration jobs will deploy a full OpenStack
setup, single node first, probably multi-node one day.

Depends-On: I4b8f7fa7a694743758ea2073b532184c587899d1
Change-Id: Id88de8eaccea9051138f03a93ff7e23135262ee7