Updated openstack/openstack Project: openstack/tempest  d95efa9be25ab6a6b6daaf9971877978713f2ea0
Add plugin docs section on configuration options

This commit adds a subsubsection to the plugin interface docs to
elaborate a bit on how to deal with configuration options. Previously,
how to use them in a plugin was implied, but nothing was explicitly
stated. This corrects that to make it clear on how to deal with
configuration.

Change-Id: Id3afc9657c5419cfba61acb3c84ccfea104d1b20