Updated openstack/openstack Project: openstack/governance  91d4b383b5f8cc8bcde943e1e287b5674c5d2eac
Moving the RefStack project from openstack-infra to Big Tent.

RefStack is a test result collection and reporting service to
support the DefCore testing process. It consumes the output of
DefCore mostly in the form of a Board-approved JSON document,
which it uses to determine whether a cloud has passed all the
tests required by the DefCore Guideline in question.

RefStack has been approved to move from stackforge to
openstack-infra [1]. However, there are increasing feedback
and general consensus of the Foundation staff is that RefStack
should be a Big Tent project.

RefStack has elected PTL and core members.  The core members
approve all code changes.  RefStack code is submitted and
reviewed using the OpenStack infrastructure [2].

RefStack has been running weekly meetings from the start of the
project. Meetings are held on Monday at 19:00 UTC and are open
to anyone that wants to attend. Our Weekly meeting agenda is
published at the meeting wiki [3].

The #refstack IRC channel on freenode is logged and the channel is
notified of such logging [4].

A RefStack server (refstack.openstack.org) will be hosted by
OpenStack Infra [5][6].

[1]https://review.openstack.org/#/c/205609/
[2]https://review.openstack.org/#/q/project:stackforge/refstack,n,z
[3]https://wiki.openstack.org/wiki/Meetings/Refstack
[4]http://eavesdrop.openstack.org/irclogs/%23refstack/
[5]https://review.openstack.org/#/c/188207/
[6[https://review.openstack.org/#/c/198869/

Change-Id: I2a9f029acc6908deb5486567eadb311b43dffbdb