Updated openstack/openstack Project: openstack/barbican  c40315291cb0269a32fada5343e147c0eae4cf49
Add Quota Enforcement API

This commit adds the quota enforcement API.  When emplemented, enforce()
will return without exception if a new resource can be created without
exceeding a quota limit.  enforce() will raise an exception if the
creation needs to be blocked due to the effective quota.  A detailed
error message is created for both the log and the response body of
the create request.

Change-Id: Iaf2a3a3e5960e84c3fe7d9d2becf1cc32b788833
Partially-implements: blueprint quota-support-on-barbican-resources