Updated openstack/openstack Project: openstack-infra/project-config  2f94bc357650671feb910c9aa26933728c61b790
Gate ceilometer on its plugin-based grenade job

Make the experimental grenade-dsvm-ceilometer voting and drop
ceilometer from the standard grenade.

The experimental job has been running well for a significant number
of tests but is in the awkward state of upgrading ceilometer twice
because ceilometer remains in core grenade. It is removed in the
depend-on patch below.

Depends-On: I09f47a8c24ac59e5db840bcbc7970bfb5d259591
Change-Id: I2517bfde5795bd98e9a89e416445f1c04c8b51a3