Updated openstack/openstack Project: openstack/manila  e48fbd8dc197365c7e3c39d38fa3f2c4739bee4b
Lazy Load Services

When the Manila share service starts, it instantiates
the Neutron network helper, which makes a call to the
Neutron service. If the service is started and the
Neutron service is unaccessible, the share service will
fail to even start.

This patch lazy loads those parameters so that they are
only loaded when needed (and presumably when the Neutron
service is up).

Change-Id: I1857c0cb6fc41afd0d01916addc1fd216070ffd7
Closes-Bug: #1481331