Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  5ce1b32e027a0e8bdafb0e90e79ce045e06d3ffc
Allow whitespace to be significant

Remove the test options suppressing ignoring differences in whitespace
outputted in the XML to allow capturing of issues where the code
incorrectly adds or removes significant whitespace.

Add a helper function to deal with the edge case of parsing XML directly
that contains non significant whitespace to avoid accidentally adding
excess whitespace when minidom outputs the resulting document.

Change-Id: I9936042cd82c204ba2b3c19f575703e33564f7fd