Updated openstack/openstack Project: openstack/nova  f47f6c740efa048d6dcf4f852a6b644d21e4d810
Move objects registration in tests directory

In order to use a NovaObject via the magic objects.Foo syntax
objects.register_all() must have been called first.  This was being done
in nova/test.py which most files in the test directory imported due to
TestCase being defined there.

Files in the test directory which provides fakes for use in testing do
not typically import nova.test and were not able to use objects.Foo to
instantiate an object.  This moves the objects registration to
tests/unit/__init__.py so that all modules in the unit tests directory
can access objects in the same way.  Since fakes are usually defined in
the unit tests dir this does not change anything for tests/functional/.

Change-Id: I7dee851cb4c37af3becf948cb55da5ee011a1e5e