Updated openstack/openstack Project: openstack/horizon  b1eab57e4da3875cfc49b106527a6108f789119a
Correct local_settings example from overwriting settings

Currently, local_settings.py.example redefines HORIZON_CONFIG
(previously defined in settings.py). This adds to confusion as to
where proper definitions and modifications should be made.
This change ensures that only modifications of HORIZON_CONFIG
are made within local_settings.py.example and defining of
HORIZON_CONFIG is left in settings.py.

-added importing of HORIZON_CONFIG from settings

-corrected portion that was redefining HORIZON_CONFIG

-left the modal_backdrop = static change that was
occuring in the redefition in local_settings to
that of a modification and commented out to serve
as an example (conforms to usual pattern)

-migrated the default for 'modal_backdrop'
into settings.py

Change-Id: I7ddeee82258f6c871ff77e1bc03115c2adbf618e
Closes-Bug: #1472687