Updated openstack/openstack Project: openstack/sahara-image-elements  6010b9a4e1a9e0912bd0b6e345fde02b9f56ca35
Setup JAVA_HOME for cloudera java distro

This patch makes jdk dependency for cloudera plugin and
setup JAVA_HOME for this plugin.
This is needed for correct work of ssl.

Note: as we have added java dependency so we need to export
DIB_JAVA_DISTRO variable for passing check in
'.../elements/java/pre-install.d/00-check-java'

Change-Id: Ic505eb059783625a4f2e452414b45202a5730caf