Updated openstack/openstack Project: openstack/oslo.versionedobjects  00c503efbb90f03d00c72dae8baeeab6ae4289fc
Allow calling obj_to_primitive() with version_manifest

When we call obj_to_primitive() on an object, we need to be able to
pass in the version_manifest so that the backport happens according
to those versions. This merely adds an optional argument which, if
present, calls the new interface.

Change-Id: I016a8061825507751168b6a59e092c9db5b0b19f