Updated openstack/openstack Project: openstack/glance-specs  f789ae898eb58cf1c6a82fbed618fdae93e8be10
Fix deactivate/reactivate error to 403 in spec

As it is, Glance correctly returns a 403 when one attempts to
deactivate (reactivate) an image that is not in active (deactivated)
status. A typo in the current spec incorrectly states that a 400
is returned in this situation. This patch corrects the typo.

No code fix is required.

Closes-Bug: #1445487

Change-Id: I7d7ed4716bffa69fa5f7f17145a1b64acc1cdbfa