Updated openstack/openstack Project: openstack/cinder  73a51dd21c1467f3889a6056b630e0b6a8d7f9c4
Cleaning up CONF.register_opts() in compute/__init__.py

This patch makes the import of oslo_config and registration
of _compute_opts consistent with all other imports and
registration statements in other files. In a future patch
that dynamically generates Cinder config options, it will now
be possible to get the options being registered in this file.

Change-Id: Ifaae49dd6d0163b5cde44df7b3ca904fb550f211