Updated openstack/openstack Project: openstack/tempest  0c641c90d4d07c3de75da9960b39e5672d9eb1da
remove test_update_host_with_extra_param

This negative tempest test assumes that input validation will happen
on parameters before platform validation of the feature. This feature
is only supported on the xenapi backend.

There is no explicit guaruntee that input validation will always trump
platform validation, and when applying the relaxed validation of Nova
v2.0 on v2.1 on a platform that will return a NotImplemented, we might
get the NotImplemented back first.

Given that this really ends up being a xen platform specific test, and
a negative test at that, it doesn't bring a lot of value to tempest
and should just be deleted.

Change-Id: I513b54eeacee1486ea11b85913ab2d21727fc5f9