Updated openstack/openstack Project: openstack-infra/os-loganalyze  c4a1cf8f463491c37157c03768db5080dc6d519f
Whitelist .log.gz files from the passthrough

Currently, files with extension .log.gz are not exempt from the
passthrough filter and will be downloaded by the browser instead
of formatted for viewing in the browser. Since some log files in
stable releases use .log for the extension, it's useful to
whitelist these files when served up for log analyzation.

Change-Id: I3d391167bbc3e3b8a4fffd368a0a16f944f548ca