Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  c7c8e12e6cd7c51fe553aeafedce331d4947a86f
Stash publisher's credentials configuration

As per now if you want to configure the stash publisher and give it
a username and password, you need to set them in clear text within
the yaml file. This pull request gives the possibility to set these
credentials within the jenkins_jobs.ini file and therefore makes it
possible to not expose them to everyone.

Change-Id: I2e5d6caefb87ded2468c7b7e015e20464ffef99e