Updated openstack/openstack Project: openstack/keystone  cad37b35eb98ed5b0819a93011ad1dda1c2f4da4
When validating a V3 token as V2, use the v3_to_v2 conversion

There is no need to use the V2 token formatter, which is a separate
code path from the v3_to_v2_token code (needed for Fernet tokens).
This patchset aligns the v2 token validation when the original token
is V3 to work with the v3_to_v2_token converter.

Change-Id: Ia88c560e89abdd937376c5c38770298ab9adcef5