Updated openstack/openstack Project: openstack/keystone  94b9a66b51d6f9794f961a32d380c2d8392aa446
Prevent exception for invalidly encoded parameters

An exception occurs in the WebOb library when it tries to decode a
parameter that is not valid UTF8. To avoid this let's wrap the section
using the WebOb code and raise a ValidationError so the user gets
proper feedback instead of a UnknownError and a server 500 response.

Change-Id: I6781770a6c5b317eaef84064ef61b9c838d22bc2
Closes-Bug: 1485694