Updated openstack/openstack Project: openstack/sahara  99d7629e9e1a6f75dbe88c40d41637ed20fca843
Ensure working dir is on driver class path for Spark/Swift

For Spark/Swift integration, we use a wrapper class to set up
the hadoop environment.  For this to succeed, the current
working directory must be on the classpath. Newer versions of
Spark have changed how the default classpath is generated, so
Sahara must ensure explicitly that the working dir will be
included.

Change-Id: I6680bf8736cada93e87821ef37de3c3b4202ead4
Close-Bug: #1486544