Updated openstack/openstack Project: openstack-infra/project-config  54623edd2c9f64a7aadbb28a37dcb5800a1145de
Set up new selenium job for Horizon

This job hands off management of the X virtual frame buffer to
Horizon. The interaction between the run-selenium.sh xvfb
management and Horizon's (necessary for dev reasons) was causing
the whole thing to not work.

Remove the gate-horizon-selenium job which wasn't running any
tests anyway, and add in a new job gate-horizon-selenium-headless.

The new job name is to ensure we aren't still running the old job
and to not be confused with the selenium job template.

Change-Id: I370880c17210c4a2d203913cdb49de1804ccf709
Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>