Updated openstack/openstack Project: openstack/neutron  9ed4be7559b1b611b35b505a786c3511dccac58f
Quota enforcement: remove locks on _dirty_tenants

This lock was used to avoid errors due to list contents changing
during iteration, but is causing issues with pymysql. This patch
proposes an alternative approach which makes the use of a lock
unnecessary.

With this change a copy of the dirty_tenants set is made before
setting the dirty bit on resources, and then the mark_dirty routine
operates on this copy. This still guaranteses operations
correctness, as all the tenants that should be marked dirty are
marked dirty before the completion of the relevant API request.

Related-Blueprint: better-quotas

Change-Id: Ib39e7089889d3f906bdc025c843128a1fa3e8797